pyproject.toml
[project]
name = "twitch-data-ingestor"
version = "0.1.0"
description = "A Twitch data ingestion service"
authors = [{ name = "Ben Ledingham" }]
dependencies = [
    "requests",
    "pytest-cov",
    "python-dotenv",
    "ruff",
]
requires-python = ">=3.10"
